Add new build system using autoconf + GNU make
This work is partially based on the ScummVM build system (which in turn
borrows from the pentagram build system), partially on the git.git build
system.

Note that it deliberately does not use GNU automake: After extensive
experiments with that, I felt that I had to spend far too much time
working around its quirks, and trying to understand the complexities
hidden behind its opaque commands.

Instead, I choose to add a Makefile which requires GNU make. While this
requires a little bit more work from us for a few things (e.g.
dependency tracking), it is much simpler to understand and hence also
easier to debug, as it doesn't try to hide the complexity behind m4
macro tricks.

Among the key benefits, it enables us to handle HPC-GAP mode in clever
ways. see the comments in Makefile.rules for details.

Also new is the way we handle the GAP version. The new approach is based
on what git.git does, and much more robust than what we have in the
current master branch.