Remove dev/DistributionUpdate/rsyncserv
These files seem to be for Frank's rsync distribution of GAP. I am
primarily removing them right now as they lead to lots of irrelevant
hits when grepping for filenames in the repository. So in the future we
can revert this commit (or rather, drop it during a rebase of the
hpc-merge branch).

However, I'd also argue that they don't belong in here either way;
but rather, they should be in a repository of their own. But perhaps
there are strong reasons why they should be in the gap repository
after all? Let's hear the arguments then!