WIP: Compiler bundled GMP if we cannot find a usable external copy
On most systems, we will want to use a system-wide version of GMP.
But some users may not have one, may not even have the necessary
privileges to install one. For others, they may want to compile
a 32b it version of GAP but only have a 64 bit version.

So, if necessary, compiler the version of GMP we bundle with our source code.

TODO:

- ensure gmp is built before linking?
- ensure 'make install' does the right thing (whatever
  that is, exactly?????)